:truck: move AdminMember model to team boundary
