+++ api/src/team/domain/models/AdminMember.js
@@ -4,7 +4,7 @@ const Joi = BaseJoi.extend(JoiDate);
 import lodash from 'lodash';
 
 import { PIX_ADMIN } from '../../../authorization/domain/constants.js';
-import { validateEntity } from '../validators/entity-validator.js';
+import { validateEntity } from '../../../shared/domain/validators/entity-validator.js';
 
 const { isNil } = lodash;
