+++ api/tests/shared/integration/infrastructure/repositories/admin-member.repository.test.js
@@ -1,7 +1,7 @@
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { AdminMemberError } from '../../../../../src/authorization/domain/errors.js';
-import { AdminMember } from '../../../../../src/shared/domain/models/AdminMember.js';
 import { adminMemberRepository } from '../../../../../src/shared/infrastructure/repositories/admin-member.repository.js';
+import { AdminMember } from '../../../../../src/team/domain/models/AdminMember.js';
 import { catchErr, databaseBuilder, expect, knex, sinon } from '../../../../test-helper.js';
 
 const { ROLES } = PIX_ADMIN;
