+++ api/tests/team/unit/domain/usecases/save-admin-member.usecase.test.js
@@ -1,7 +1,7 @@
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
-import { AdminMember } from '../../../../../src/shared/domain/models/AdminMember.js';
 import { AlreadyExistingAdminMemberError } from '../../../../../src/team/domain/errors.js';
+import { AdminMember } from '../../../../../src/team/domain/models/AdminMember.js';
 import { saveAdminMember } from '../../../../../src/team/domain/usecases/save-admin-member.usecase.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../../test-helper.js';
